Library to maninpulate images, with what I need at the moment:
various types of scaling (by specifying only width, only height
and proportionally resizing the other dimension, constrained by
a bounding box and with white bands on top/bottom or at sides)
Further it can read and save a JPEG. The way it is built is object
oriented and can manipulate one image a a time. The same object
can be used to manipulate as many images as needed by calling the
read method multiple times.